+++ plugin/oneall.rock-arch/RockLens.qml
@@ -40,7 +40,7 @@ Panel {
   property bool onboardingAutomaticUpdates: false
   property bool updateManaged: false
   property string updateState: "idle"
-  property string currentVersion: "0.24.0"
+  property string currentVersion: "0.24.1"
   property string availableVersion: ""
   property string updateLastCheckedAt: ""
   property string updateLastUpdatedAt: ""